Service2/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.2
django-cors-headers==4.5.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
grpcio==1.67.0
grpcio-tools==1.67.0
protobuf==5.28.2
pycparser==2.22
PyJWT==2.9.0
redis==5.1.1
sqlparse==0.5.1
uvicorn==0.32.0

